pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-firewall"
version = "0.1.0"
description = "The open-source security gateway for AI agents. Policy enforcement, threat detection, and compliance-ready audit logging for MCP."
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.11"
authors = [
    { name = "Robert Ressl", email = "rr@canus.ch" },
]
keywords = [
    "mcp",
    "security",
    "firewall",
    "ai-agents",
    "gateway",
    "compliance",
    "dora",
    "audit",
    "policy",
    "model-context-protocol",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Networking :: Firewalls",
]
dependencies = [
    "mcp>=1.0",
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "cryptography>=42.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "websockets>=12.0",
]

[project.optional-dependencies]
opa = ["regopy>=0.3"]
pdf = ["weasyprint>=61.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
mcp-firewall = "mcp_firewall.cli:main"

[project.urls]
Homepage = "https://github.com/ressl/mcp-firewall"
Documentation = "https://github.com/ressl/mcp-firewall/tree/main/docs"
Repository = "https://github.com/ressl/mcp-firewall"
Issues = "https://github.com/ressl/mcp-firewall/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
